Изменить расположение файлов и добавить файл базового view
