--- src/basePage/basePage.handlebars
@@ -1,9 +1,0 @@
-{{> leftMenu}}
-<div class="main">
-    <div className="main__container">
-        {{> cap}}
-        {{> containerDesk}}
-    </div>
-</div>
-{{> rightMenu}}
-{{> settings}}
