+++ src/components/containerDesk/containerDesk.handlebars
@@ -1,7 +1,7 @@
 <div class="main__content">
     {{#each boards}}
         {{>desk}}
-{{/each}}
+    {{/each}}
     <div class="desk deskButton">
         <div class="desk__title">Доска для важных дел</div>
         <div class="desk__description">
