+++ src/components/leftMenu/leftMenu.handlebars
@@ -1,4 +1,4 @@
-<header class="header">
+<header class="header {{#if pageStatus.isLeftMenu}}header_open{{/if}}">
     <div class="header__logo">
         <svg width="30" height="32" viewBox="0 0 30 32" fill="none" xmlns="http://www.w3.org/2000/svg">
             <path d="M8.10705 17.9027C7.45515 17.5226 6.62098 17.7472 6.24387 18.4042C5.86677 19.0613 6.08953 19.9021 6.74144 20.2822C8.03775 21.038 8.99514 22.3146 9.33455 23.8222C9.50121 24.5625 10.2317 25.0264 10.9661 24.8585C11.7006 24.6905 12.1609 23.9542 11.9942 23.214C11.483 20.9434 10.0456 19.033 8.10705 17.9027Z" fill="#4277FA"/>
