+++ src/views/basePage/basePage.handlebars
@@ -1,0 +1,9 @@
+{{> leftMenu}}
+<div class="main {{#if pageStatus.isLeftMenu}}menu-open{{/if}} {{#unless pageStatus.isRightMenu}}active-tasks-open{{/unless}}">
+    <div class="main__container">
+        {{> cap}}
+        {{> containerDesk}}
+    </div>
+</div>
+{{> rightMenu}}
+{{> settings}}
