+++ src/views/baseView.js
@@ -1,0 +1,14 @@
+'use strict';
+
+class BaseView {
+	/**
+	 * @constructor
+	 */
+	constructor() {
+
+	}
+
+	render() {
+
+	}
+}
