+++ src/views/signupPage/signupPage.js
@@ -1,13 +1,14 @@
 'use strict';
 
-import {validateLoginPage, validateSignUpPage} from '../modules/validation.js';
-import * as render from './loginPage.templ.js';
-import {deleteListeners} from '../modules/deleteEventListeners.js';
+import * as render from './signupPage.templ.js';
+import {validateLoginPage, validateSignUpPage} from '../../modules/validation.js';
+import {addPrompt, deletePrompt} from '../../modules/prompt.js';
+import {deleteListeners} from '../../modules/deleteEventListeners.js';
 
 /**
- * Функция, осуществляющая рендер страницы логина.
+ * Функция, осуществляющая рендер страницы регистрации.
  */
-export function loginPageRender() {
+export function signupPageRender() {
 	/* Удаляем обработчики событий для всех используемых элементов */
 	deleteListeners();
 	/* Регистрация всех компонентов для страницы */
@@ -19,12 +20,9 @@ export function loginPageRender() {
 	Handlebars.registerPartial('input', Handlebars.templates['input']);
 	Handlebars.registerPartial('logo', Handlebars.templates['logo']);
 
-	/* Смена урла в адресной строке */
-	window.history.pushState('', '', 'http://localhost:3000/login');
-
 	/* Рендер шаблона с входными данными */
-	const loginPage = Handlebars.templates.loginPage;
-	const html = loginPage({});
+	const signupPage = Handlebars.templates.signupPage;
+	const html = signupPage({});
 
 	/* Создание контейнера для вставки в DOM */
 	document.body.innerHTML = '';
@@ -33,7 +31,14 @@ export function loginPageRender() {
 	document.body.appendChild(container);
 	document.getElementsByClassName('container')[0].innerHTML += html;
 
-	/* Навешивание обработчика валидации данных для формы логина */
+	/* Навешивание обработчика валидации данных для формы регистрации */
 	const form = document.getElementById('input_form');
-	form.onsubmit = validateLoginPage;
+	form.onsubmit = validateSignUpPage;
+
+	/* Добавление подсказки при вводе пароля */
+	const inputPas = document.getElementById('input_pass');
+	inputPas.addEventListener('focus', addPrompt);
+
+	/* Удаление подсказки при снятии фокуса */
+	inputPas.addEventListener('blur', deletePrompt);
 }
